// shopping-list/src/mobile/src/app/clients/api.ts
export class GetListByIdItem implements IGetListByIdItem {
    itemId!: number;
    productName?: string | undefined;
    alreadyBought!: boolean;

    constructor(data?: IGetListByIdItem) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.itemId = data["itemId"];
            this.productName = data["productName"];
            this.alreadyBought = data["alreadyBought"];
        }
    }

    static fromJS(data: any): GetListByIdItem {
        data = typeof data === 'object' ? data : {};
        let result = new GetListByIdItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["itemId"] = this.itemId;
        data["productName"] = this.productName;
        data["alreadyBought"] = this.alreadyBought;
        return data; 
    }

    clone(): GetListByIdItem {
        const json = this.toJSON();
        let result = new GetListByIdItem();
        result.init(json);
        return result;
    }
}

export interface IGetListByIdItem {
    itemId: number;
    productName?: string | undefined;
    alreadyBought: boolean;
}
// ...
export class SetAsAlreadyBoughtCommand implements ISetAsAlreadyBoughtCommand {
    shoppingListItemId!: number;
    alreadyBought!: boolean;

    constructor(data?: ISetAsAlreadyBoughtCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.shoppingListItemId = data["shoppingListItemId"];
            this.alreadyBought = data["alreadyBought"];
        }
    }

    static fromJS(data: any): SetAsAlreadyBoughtCommand {
        data = typeof data === 'object' ? data : {};
        let result = new SetAsAlreadyBoughtCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["shoppingListItemId"] = this.shoppingListItemId;
        data["alreadyBought"] = this.alreadyBought;
        return data; 
    }

    clone(): SetAsAlreadyBoughtCommand {
        const json = this.toJSON();
        let result = new SetAsAlreadyBoughtCommand();
        result.init(json);
        return result;
    }
}

export interface ISetAsAlreadyBoughtCommand {
    shoppingListItemId: number;
    alreadyBought: boolean;
}